# doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Namespace API",
    "title": "Virtual Security Module"
  },
  "basePath": "/",
  "paths": {
...
    "/namespaces": {
      "post": {
        "description": "Creates a new namespace",
        "tags": [
          "namespaces"
        ],
        "operationId": "CreateNamespace",
        "parameters": [
          {
            "name": "NamespaceEntry",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/NamespaceEntry"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/NamespaceCreationResponse"
          }
        }
      }
    },
...
    "/secrets": {
      "post": {
        "description": "Creates a new secret",
        "tags": [
          "secrets"
        ],
        "operationId": "CreateSecret",
        "parameters": [
          {
            "name": "SecretEntry",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SecretEntry"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/SecretCreationResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "NamespaceEntry": {
      "type": "object",
      "properties": {
        "childPaths": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "ChildPaths"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "path": {
          "type": "string",
          "x-go-name": "Path"
        },
        "roleLabels": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "RoleLabels"
        }
      },
      "x-go-package": "github.com/vmware/virtual-security-module/model"
    },
...
    "SecretEntry": {
      "type": "object",
      "properties": {
        "expirationTime": {
          "x-go-name": "ExpirationTime"
        },
        "id": {
          "type": "string",
          "x-go-name": "Id"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "secretData": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint8"
          },
          "x-go-name": "SecretData"
        }
      },
      "x-go-package": "github.com/vmware/virtual-security-module/model"
    },
...
  },
  "responses": {
...
    "NamespaceCreationResponse": {
      "schema": {
        "type": "object",
        "properties": {
          "NamespaceId": {
            "type": "string"
          }
        }
      }
    },
...
    "SecretCreationResponse": {
      "schema": {
        "type": "object",
        "properties": {
          "SecretId": {
            "type": "string"
          }
        }
      }
    },
...
  }
}